+++ lib/utils.ts
@@ -1,7 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-import { clsx, type ClassValue } from "clsx";
-import { twMerge } from "tailwind-merge";
-
-export function cn(...inputs: ClassValue[]) {
-  return twMerge(clsx(inputs));
-}