+++ postcss.config.mjs
@@ -1,9 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-const config = {
-  plugins: {
-    tailwindcss: {},
-    autoprefixer: {}
-  }
-};
-
-export default config;